add awful realization of formatting alternate fonts
Why is it awful? Because a lot of code is duplicated. And the code is
badly designed. As a result, we have one bug, which i don't know how
to fix (if you know - write me at email):

.BI [ "--init, -i"]
      ^ start italic font style
                  ^ must start bold, but still using italic

But I couldn't leave it as it is and have made a dirty trick:
`
- match: '"[^\s]'
  scope: invalid.illegal
  pop: true`
`

If the '"', which is supposed to end the string, has a not white-space as
following character, then I scoping it like illegal syntax. Anyway, it's not
correct to write this way, put a white-space after '"' :)